backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.11.9",
    "fastapi>=0.129.0",
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.6",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=14.1",
]
